requirements.txt
streamlit
openai
openpyxl
pandas
pillow
python-docx
PyPDF2
pillow-heif